requirements.txt
typer
tabulate
gitpython
requests
trex